fixed 5th

+++ OpenGL-HW5/Shader.cpp
@@ -95,7 +95,7 @@ GLuint ShaderProgramCreateFromFile(const char *vsName, const char *fsName)
 
 		// добавим к коду вершинного шейдера параметры
 		const GLchar *source[2] = {(const GLchar*)vertexShaderDefines, (const GLchar*)shaderSource};
-		const GLint  length[2] = {sizeof(vertexShaderDefines) - 1, sourceLength};
+		const GLint  length[2] = {sizeof(vertexShaderDefines) - 1, (int)sourceLength};
 
 		// зададим шейдеру исходный код и скомпилируем его
 		glShaderSource(shader, 2, source, length);
@@ -140,7 +140,7 @@ GLuint ShaderProgramCreateFromFile(const char *vsName, const char *fsName)
 
 		// добавим к коду фрагментного шейдера параметры
 		const GLchar *source[2] = {(const GLchar*)fragmentShaderDefines, (const GLchar*)shaderSource};
-		const GLint  length[2] = {sizeof(fragmentShaderDefines) - 1, sourceLength};
+		const GLint  length[2] = {sizeof(fragmentShaderDefines) - 1, (int)sourceLength};
 
 		// зададим шейдеру исходный код и скомпилируем его
 		glShaderSource(shader, 2, source, length);
